UIWRKFLOW-2: Create a Search Panel in the UI.
This creates much of the functionality and tooling needed.
However, this does not have all of the needed functionality.
The search box and buttons still need work.

There are problems with some of the dependencies and this temporarily requires that the `"@folio/stripes-types"` be added to the `resolutions` section in the `package.json` file.
This will need to be removed in the future once upstream dependencies have the issue resolved.

This does not do much with CQL support not yet fully implemented.