UIWRKFLOW-3: Switch to and use the search query end point.
This adds a path parameter to allow for custom paths to be used.

This removes the no longer needed `HOME_PAGE_SHORT` constant.
A `PATH` map is used to contain the search paths for each view.

The `query` is required and so use a `typeof` check so that an empty string may also be passed.
This allows the `buildListUrl` to be used on more than just a search path that requires `query`.