UIWRKFLOW-3: Simplify browser request, remove not yet used code.
The search end point does not supports ids nor does it support filters.
Remove this code for the time being.

The CQL query is the filter and there may need to be further design considerations regarding how to implement CQL query generation from the UI perspective.
The current design just provides an empty CQL query.

The rest of the filter code is left in place to either be amended, replaced, or deleted at some future date.