UIWRKFLOW-3: The workflow column name is `versionTag` and not `version`.
I cannot simply have the properties be `VERSION` but the detailed value be `versionTag` as this somehow causes compilation problems.
Refactor every `VERSION` to `VERSIONTAG`.
Refactor every `version` to `versionTag`.

This update the `en.json` and leaves the other languages alone.
This is done because I expect the automated processes will create the language translations for us like it previously has done.