UIWRKFLOW-3: Create Results in the UI.
The table shows a list with the Name, Description, Version, and Active/Inactive.

The panel is updated and the searching is working.
The search also includes the ID, which is not displayed in the table.

The `searchSettings` is added but is not used.
The CQL needs to be sanitized for CQL purposes.
The values are auto-URL-encoded.
More research is needed for this.

The settings have some initial tooling setup.
A new permission is added for allowing to make changes to the settings at a non-granular level.

The `WorkflowSettingsForm` is incomplete and produces an empty form.
The following code was the initial attempt but there are compile problems with the `TextField` not being supported there for some reason:
```ts
      <Pane
        defaultWidth="fill"
        fluidContentWidth
        id="pane-workflow-settings-display"
        lastMenu={lastMenu}
        paneTitle={label}
      >
        <Field
          component={TextField}
          fullWidth
          id="workflowsPaginationSize"
          name="workflowsPaginationSize"
          label={fieldLabel}
          type="number"
        />
        <Layout className="padding-bottom-gutter padding-top-gutter">
          <FormattedMessage id="ui-workflow.settings.workflows.paginationSize.description" />
        </Layout>
      </Pane>
```