UIWRKFLOW-7: Add Import pane, add import file upload, and address code consistency.
This is the first part of a series of commits for implementing the FWZ import feature.

This does not implement the network communication and response processing.
This adds some drafted out "callouts" but the implementation is expected to change once the network communication and response processing is written.

Add `@folio/stripes-data-transfer-components` to provide the file upload drop box.
Just adding this is not enough.
The module cannot be found and custom `typings` must be created to give this dependency a type.
The `typeRoorts` is added to the `tsconfig.json` to address this.
The directory `typings` now exists at the top level of the project.
The `typings` directory contains the module declaration for getting this dependency to work.

The action menu for the Create view now has a button for opening up the Import pane.
The action menu cannot support this as originally planned due to callbacks not being called while in the drop down and also because of the drop down auto-closing when the mouse navigates off the page to select a file to drop.

The filter pane title is changed to be more customizable so that it can become more accurate.
The use of the word 'search' is currently maintained as no decision has yet to be made regarding what it should be named.

Several of the exports in several files, mostly `index.ts` files, are lacking semicolons at the end of lines.
This is rather inconsistent.
Change the code to have semicolons in all cases that I was able to find.

The "on*" default callbacks are not being found.
This is possibly related to JavaScript dependency being brought into react typescript.
Manually implementing the "on*" callbacks, such as `onDragEnter` and `onDragLeave` resolves this issue.

This adds an Import pane but also updates the existing Item Record Detail pane for consistency reasons.
Both panes are now passed a `view` property.
A new generic pane interface is created that both panes now either use or extend.

Remove unused imports that have been discovered.

I found the code to be rather inconsistent in the use of single and double quotes.
I arbitrarily chose single quotes to make the structure consistent.
Being an arbitrary decision, there is a notable chance that this may change in the future to double quotes.
The translations still use double quotes to be consistent with other projects.
The auto-translator uses double quotes to generate other language files from the `en.json`.

The class names such as `useDetailPaneSelect` are rather confusing.
I removed the word `select` from the case and updated the affected code.
Similar language improvements have been made to make the code easier to read.
I only changed things related to what I needed to work on in this regard.

Expose the setShowDetail to the "use*" pane methods to allow for changing the value.
This is needed in particular by the FWZ importer.

The `IListProperties` is a generic list property interface.
Remove the explicit `IItemRecordDetail` type.
Having that is problematic for when the `detailPaneSelect` is set to the import pane function.