UIWRKFLOW-7: Implement the network communication and response processing of the FWZ importer.
Expose `stripes` to all more of the project via a single variable passed along.
The interfaces are updated to support this.

The Dropdown menu button for the import now does a permissions check via `stripes`.
If not authorized, then the button is disabled.

The `react-query` is in direct use but the dependency doesn't explicitly exist.
Explicitly add the `react-query` dependency.