UIWRKFLOW-7: The typings directory and file is still required.
This is not obvious is that there is some sort of caching going on.
Rebuilding wasn't enough to expose that I needed to add the typings back.
I happened to reboot my machine and then the problem that the `typings` was added for re-appeared.

This adds the `typings` directory and contents back.