UIWRKFLOW-19: The DropdownMenu requires that the open be specified when onToggle is specified.
This is a continuation of the commit 248b82caded69a706845fc740dea52b379b31a9c.
The first approach set the open attribute value to the `open` variable.
This allowed for the module to compile but the drown downs became non-functional.

Simply adding the presence of the open attribute with no value assigned appears to allow for the drop down to work as expected.