UIWRKFLOW-20: Clean up the design to make extending and debugging easier.
Simplify the `ItemREcordDetailPane` to better allow for modification in a more object oriented manner.
This simplifies the markup code and removes a lot of inline logic.

This will help make debugger the problems mentioned by UIWRKFLOW-20 easier to resolve.

The language file is also simplified and the alphabetic ordering is fixed.