UIWRKFLOW-13: Change workflow detail pane to be full screen.
Utilize a `Layer` to provide Workflow details as full screen.
The other panes will be wrapped inside their own `Layer`.

The `Layer` `isOpen` property can be used to avoid the need to return NULL.

The `ErrorBoundary` is moved inside the `Pane` so that the error is more cleanly printed.
If this becomes a problem then the error will need to be moved outside.
There is still an outermost error message to catch any errors above that.