UIWRKFLOW-14: Workflow Details View is now a three-column view of General, Graph, and Details.
A new view called `ItemRecordView` is created to display the three panes:
- `ItemRecordGeneralPane`: The General Pane will house the general metadata.
- `ItemRecordGraphPane`: The Graph Pane now shows the details title and will show the visual graph structure.
- `ItemRecordDetailPane`: This remains with a few modifications and some functionality extracted into the other views.

The close button is moved from the `ItemRecordDetailPane` into the `ItemRecordGeneralPane` so that the close button is on the far left side.
None of the three panes can have their widths adjusted at this time.
The `ErrorBoundary` is moved inside the `Pane` so that the errors display in the pane rather than outside of it.
This makes the error printing cleaner and easier to read but this also has not been well tested.

Some of the interfaces and hooks were renamed to make more sense with the new structure.
- The `useItemRecordDetailPane` hook is now `useItemRecordData`.
- The `IItemRecordDetailPane` interface is now `IItemRecordData`.

In many relevant areas the `FormattedMessage` usage is converted into `t()` usage.
Simplify the style practice by avoiding unnecessary parenthesis when returning markup.

This does not populate the panes as requested by the issue.

This adds `identity-obj-proxy` dependency to address unit test problems.
I had too many problems with the unit test system that I had to abandon my attempted tests.
I will have to come back and write the tests at a later time when I am able to spend the necessary time to fix the unit test design problems.