UIWRKFLOW-17: Selecting a Node graph of the Workflow Detail should appear in the rightmost column.
## Additional Issues

The issue request includes "Also finish the implementation of the rendering of the details.".
However, this turned out to be a rather large request and there will instead be supplement issue created to represent additional work that is needed.

The additional work in short is:
1. Recursive processing of Nodes.
2. Proper display of Code (Ruby, JavaScript, etc...).
3. Proper display of SQL.
4. Proper display of large plain text and HTML markup.
5. Proper display of URI (URL, etc...).
6. Embedded types and embedded variables such as Input Variables and Output Variable.
7. Proper and safe display of passwords.

For (1), the work to do this was originally attempted.
This turned out to be troublesome and time consuming.
I backed off on the work with plans to resume (now as as separate issue).

For (2), (3), and (4), the code needs to be rendered in a friendly way.
The current representation is raw in some cases.
In cases of code, the code is JSON encoded and needs to be decoded (and is also wrapped in quotes).

For (5), a URI (URL, etc...) should render as a URL that can be clicked, etc...

For (6), this is closely related to (1) and there may be code reuse.

For (7), there will likely need to be some sort of access control decisions and not just showing and hiding of the display of the password.
The current behavior is just to not display it and instead print `(hidden)`.

The "For lengthy text, such as is seen for the script data (JavaScript, Ruby, SQL, etc..), we need to come up with a generic solution that looks “all-right”." requirements is now expected to be resolved by (2), (3) and (4).
The "Example implementation idea: For known types like Script or SQL delegates, perform a length check." requirement is also now expected to be resolved by (2), (3) and (4).

## Language Changes

The language file has been restructured to simplify it.
Furthermore, the capitalization structure is changed to be like a sentence.
While this capitalization practice is bad grammar and bad practice for a title, buttons, labels, etc..., this is the FOLIO standard way of doing things.

see: https://ux.folio.org/docs/guidelines/language-rules/text-case-rules/

## Major Refactoring

Much of the code structure is refactored to simplify the implementation, reduce redundancy, and increase maintainability.
This has been deemed necessary to make the more modular structure used for the Node Details to work well.

A lot of documentation comments have been added to help improve the readability.

The use of more gneric `props?: any` as a parameter to many of the `React.FC` functions has allowed for easier use of the views.
The data can now be passed along using `{ ...props }`, thereby reducing the need to change those files if the property structure ever changes.

## Problems and Work-arounds

The Panes have been configured in such a way that they work with the current design and structure.
Much of this is a work-around to what is actually desired.
In some cases the Panes do not work right or work as expected when sane defaults of just setting a percentage width.
This appears to work well for the Import Pane but for unknown reasons (the same setup) on other Panes do not work in a sane manner.

Addressing these problems is pushed off to he addressed in another ticket that must be created.
Investigation and resolution of the problems is needed but is considered out of scope for this issue.

## Visual Layout Decisions

I have been uncertain as to how the display of the details should be done.
I opted to go for a random structure order based on the order of the parameters in the back end.

An exception to this are the following:
1. The main settings of the Node (`name`, `description`, `id`, etc...).
2. The Variables, such as Input Variables and Output Variable.
3. The Nodes (see `Subprocess` type).

I would love for the "General Pane" and the "Graph Pane" (Displays the Workflow Node graph) to be at the following ration:
- 25% for "General Pane".
- 75% for "Graph Pane".

Due to the bugs mentioned in the "Problems and Work-arounds" section, these are instead at 50% each.

When the "Detail Pane" is oppened, this could then be:
- 25% for "General Pane".
- 25% for "Graph Pane".
- 50% for "Detail Pane".

But again due to the problems, this is currently set to about 33% for all three.

A new CSS file is added that uses `oklch`.
I am not very familiar with this and the color structure is not easy to read and understand.
However, the percentage to make things darker works well and is supposed to be more accessible than the older types (such as rgba).
I am forced to use `!important` in one case due to a problem with the `Card` CSS from `stripes-components` not being easily overrideable.
It would be great to find a way to not use `!important` as its use is bad practice.

## Object Structure Decisions

The properties are now changed to be more like Object-Oriented-Programming in that these structures are used like self-contained Objects with their respective goal.
For example, consider `BooleanItemValue` and `StringItemValue`.
The `BooleanItemValue` is dedicated for taking the Node property and displaying it as a boolean value converted to a string for a user to view and read.
This handles the different types of Booleans so that the caller does not have to care.

For the `StringItemValue` takes the Node property and displays it directly as a string.
Read the documentation comments for details on the ItemValue types but, for the most part, the name of the Object should be self-explanatory.

In all cases the `empty`, `label`, and `value` are used.
Where `empty` if specified will suppress an empty value.
Where `label` is the ID of the translation key that gets mapped to the label for a particular language.
Where `value` is the value (or array of values) being processed by that Object.

In some specific cases such as with the `BooleanItemValue` the possible values are known and so the value is passed through the translation mapping.
In almost all other cases the values are often rendered without translation mapping.

Several of the Objects are currently functionally identical to the `StringItemValue`.
This is because the detailed changes to them are to be resolved in new issues soon to be created.

Other Objects always return `null`.
This is because they have no special properties to display.
It would be easier to not display them at all.
However, I decided that it would make the code more maintainable by having them there.
Should properties be added to them then that file only need to be slightly modified.
This design decision also helps show that the type that the Object represents is not accidentally omitted.

The Objects themselves also determine if they should or should not render based on the `deserializeAs` property.
This simplifies the code that the caller has to write.

## Behavioral Design

The Node graphs now have a Title, a Description, and a Type.
The Type is on the top-right side of the Card that display the Node graph.

This introduces a new click listener for each Node graph.
The Details Pane is opened when any Node graph is clicked on.
The Details Pane has its own dedicated close button.

There is currently no recursion but there needs to be.
This is expected to be resolved by another soon to be created issue as described in the "Additional Issues" section.