UIWRKFLOW-23: The Detail Pane View should display code that "looks all right" (including SQL and Markup).
Provide a box that display the unserialized code.
I opted to add styling to the box to make the code block more distinct as code.
This look and feel could potentially be undersirable and will require product owner review.

An arbtrary 256 character limit (`BRIEF_CODE_SIZE`) is imposed on the (unserialized) code on the side pane.
If the (unserialized) code exceeds this value, then ellipses are added (U+2026) `…` as well as a button and an associated modal.
The associated modal will open up on button click to display the entire (unserialized) code without truncation.

The different `ItemValues` needing this functionality are now merged into a single `ItemValue` where they now all expect to utilize the `format` attribute to designate the type.

Several types, such as `application/html`, are not JSON encoded and must not be decoded.
To support this a `raw` attribute is provided that handles this.

The `JSON.parse()` is fragile and cannot handle an empty string.
Perform additional checks to make sure `JSON.parse()` is not performed on an empty string so that a pointless exception does not get thrown.

The `aria-label` is added for accessiblity.
I added the text that feels correct but I don't really know how it should best be worded.

I have `closeOnBackgroundClick` added to the `Modal` but that doesn't appear to work.

Use the Unicode character U+2022 (Bullet) `•` to separate title/label groups based on this FOLIO style guide documentation:
- https://ux.folio.org/docs/guidelines/language-rules/pane-header-language/