UIWRKFLOW-18: The action dropdown in the workflow detail view includes a delete workflow option.
The delete workflow action buton does not make sense on the detail pane that has the selected Node Graph data.
Move the Action button onto the Workflow Item General pane.

Add a delete button.
This implements a generic click handler that can help simplify some of the tooling.
Other parts of the project may be updated for consistency later.

I wanted to use a constant for the delete path but having the id in the middle makes it less practical.
I need to do further research to see if I can find a way to use a constant in this way.
Finding a way to do the constant as a template string and then later substitute the variables is therefore considered out of scope at this time.

A callback is used to display errors.

This is based on the `ImportPane`.
Some of the shared functionality between these two are moved into common structures.