UIWRKFLOW-18: Ensure the delete modal has a confirmation modal.
This implements a `useModal` hook in a generic manner to provide the management of the modal state.
I can see tech-debt to improve the rest of the project to use the same design and perhaps same modal where needed.

This follows along the lines of the FOLIO recommendations here:
- https://ux.folio.org/docs/guidelines/language-rules/confirmation-message-language/#media-13808

The pane's `onClose()` methid is renamed to `closeWorkflowPane()` because of the multiple close and hide functions now present.