UIWRKFLOW-28: Re-organizae source structure.
Add sub-directories for organizing the different hooks into groups.
Add sub-directories for organizing the different components into groups.
Add sub-directories for organizing the different utilities into groups.
Remove the plural (the trailing `s`) from the interface sub-directory names.

Get rid of some of the `useState()` calls inside of the components and implement them as hooks.

Remove redundant hooks.

The `useCallbacks()` are provided within the hooks to automate the caching of the functions.
See this documentation for reasoning:
- https://react.dev/reference/react/useCallback#optimizing-a-custom-hook