UIWRKFLOW-28: Add several unit tests and fix any observed problems.
This only implements the most basic parts of the unit tests for the tests that are added in this commit.
More thorough tests are still needed.

Add `QueryCLientProvider` (and related) to get the `useWorkflowList` test working.
The documentation from the referenced link below is not entirely accurate and does not work as described.
I had to switch the test to a `.tsx` file and then make the `{ children }` optional.
I could not specify `({ children?: any })` as it complained.
I instead specified `(props?: any)` and then referenced `{props?.children}` to work-around this problem.
A `createWrapper()` outside of the test is created so that if and when there are more than one test, then the creation of the wrapper will require less code for each test.

see: https://tanstack.com/query/v4/docs/framework/react/guides/testing

The runtime does not complain about the state being set to `null` but the unit tests do.
Change the `null` to `undefined` to make the unit tests happy.

The unit tests complain about having `<IListResponse<IItemRecord[]>, Error>` on the `useQuery()`.
Remove that.
Some documentation states that adding the types in this way used to be the recommended practice but no longer is.

see: https://tkdodo.eu/blog/react-query-and-type-script#generics