UIWRKFLOW-23: Ensure when there is no code that the '-' is showed.
The standard FOLIO way of showing empty content in a `<KeyValue>` is to provide a dash.
This is automatic if the `<KeyValue>` has no children.
This is achieved through the use of `undefined`.

Rewrite the affected code to ensure that if the string is empty then the value is set to `undefined` rather than an empty string.

The wrapping `<div>` must also not be present when the `text` is `undefined`.
Only render the `<div>` inside the `<KeyValue>` by passing `undefined` as the `markup` when the `text` is also `undefined`.
When the `text` is defined, then the `markup` will wrap the `text` in a `<div>`.