UIWRKFLOW-18: Refresh Workflow list table upon deleting a Workflow.
Restructure part of the project to make the list information available to be triggered on the delete.
This exposes the `refretch` callback from the React `useQuery()`.

Clean up unused parameters and add some comments.
Shorten or simplify the name of some other variables and parameters.

The name of the Workflow list table should be called `WorkflowListTable` rather than `MainListTable`.