UIWRKFLOW-26: Display Embedded Variables.
The InputVariables and OutputVariable are both types of an Embedded Variable.
These are displayed as a table that lists all of the properties.

Restructure the details sources, relocating them under the components directory.
Implement and use the `ItemRecordNodeSet` to reduce some of the repition in many of the item record node components.