UIWRKFLOW-31: Enable routing support for Details, loading in the specific Workflow, and re-organizing accordingly.
Rework the Views so that the routing can work as expected.
Loading from the route should load the specific Workflow.

The back end is not provided the correct information due to the use of Spring Data Rest.
This requires a change in the back end to provide that actual Workflow on response rather than the incomplete one that Spring Data Rest provides.

The navigation structure has been improved to work more correctly.
An initial layout is indeed required to get things to work.

There needs to be more work regarding when to load the list and when not to.
The short term solution here is to expose the list in the newly added `MainView` that allows for the `BrowseView, `CreateView`, and `DetailView` to access the list and trigger a refresh as needed.
This likely can be improved upon.