UIWRKFLOW-39: Get things working at least under Quesnelia.
This gets the project closer to working under Poppy.

Poppy is not currently working due to dependency problems.
The Poppy build might have to have changes to the dependencies which might break other projects.

The ui-workflow project appears to require at minimum the following:
```
  "@folio/stripes-acq-components": "^5.1.0"
  "@folio/stripes-smart-components": "^9.1.0"
  "ky": "^0.31.0"
```

Remove the dependencies and resolutions that are not needed anymore.

There are dependency poblems with `ConfigManager` in Poppy.
Change the use of `UnconnectedConfigManager` to instead use `ConnectedConfigManager`.
This might help make the dependency navigation less confusing about avoiding renaming an import.
The variable is instead renamed rather than the import.

Fix the `tsconfig.json` tabbing.
Fix the include and exclude dependencies in `tsconfig.json` to reduce what is being brought in.