Get linter working and passing.
This warning message is currently printed:
```
WARNING: You are currently running a version of TypeScript which is not officially supported by @typescript-eslint/typescript-estree.
You may find that it works just fine, or you may not.
SUPPORTED TYPESCRIPT VERSIONS: >=4.3.5 <5.4.0
YOUR TYPESCRIPT VERSION: 5.6.3
```

Some of the linting does not work as a result of this discrepency and those files have been added to the `.eslintignore` file.
The latest version of eslint cannot be upgraded to yet and has problems.
see: https://folio-org.atlassian.net/browse/ESCONF-45

The versions used are the version identified to work without requiring a migration of the Eslint files.

Package dependencies have been added to ensure the `yarn lint` works.

There are a lot of problems with the linter where it makes mistakes or provides highly problematic requirements.
These have been turned off via the `package.json` `rules` for the linter.

The linter doesn't like `&&` and `||` intermixed.
Move boolean "truthy" check into its a utility function called `isFolioTruthy`.

The use of signle quotes should be on JavaScript and TypeScript strings.
The use of double quotes should be on markup tag attribute strings.

There are a lot of places where the TSX is in use and the function returns while defining variables.
This causes "early return" problems that is not supposed to be allowed.
Change the logic and flow to have the constants defined before returning.
It looks like nested functions and similar are not affected by this requirement.

The linter wants the imports to use the shortest path possible.
The linter want the imports to omit trialing slashes.

The `===` should be used instead of `==` in most places.
The `const` should be used whenever possible over `let` (and `var`).

Fix numerous style inconsistencies, including tabbing issues.

Add missing semicolons.

Pass `filterCount` as a direct value.
The linter wants that to be a `const` instead of a `let`, but at that point passing a literal seems to make more sense.

There are a lot of cases where the linter designates how the `useCallback` variables should or should not be passed.
Make all the recommended changes.

The use of dynamically genereating the `filterConfig` appears to make react unhappy according to the linter.
Re-design the code to have the `filterConfig` definition be performed once.

Remove unused variables.

Collapse string appends into the backtick expansion strings.

Move the `layouts` definition outside of the Component.
The `layouts` needs to only be defined once and defining it every Component seems like a waste.

The `deleteControl` from `ItemRecordGeneralPane` needs to be declared before `deleteModal()` uses it according to the linter.
Make this so, but the `useClickControl()` must be defined after the `deleteModal()` is defined.